+++ packages/patternfly-4/react-charts/src/components/ChartTheme/styles/scatter-styles.ts
@@ -1,0 +1,11 @@
+/* eslint-disable camelcase */
+import {
+  chart_scatter_active_size,
+  chart_scatter_size
+} from '@patternfly/react-tokens';
+
+// Donut styles
+export const ScatterStyles = {
+  activeSize: chart_scatter_active_size.value,
+  size: chart_scatter_size.value
+};