+++ packages/patternfly-4/react-core/src/components/LoginPage/LoginForm.tsx
@@ -101,7 +101,7 @@ export const LoginForm: React.FunctionComponent<LoginFormProps> = ({
         <Checkbox
           id="pf-login-remember-me-id"
           label={rememberMeLabel}
-          checked={isRememberMeChecked}
+          isChecked={isRememberMeChecked}
           onChange={onChangeRememberMe}
         />
       </FormGroup>