+++ packages/patternfly-4/react-integration/demo-app-ts/src/components/demos/ExpandableDemo/ExpandableDemo.tsx
@@ -31,6 +31,8 @@ export class ExpandableDemo extends React.Component<null, ExpandableState> {
         <br />
         <h1> Uncontrolled Expandable Example: </h1>
         <Expandable toggleText="Show More">This content is visible only when the component is expanded.</Expandable>
+        <h1> Uncontrolled Dynamic Expandable Example: </h1>
+        <Expandable toggleTextExpanded="Show Less" toggleTextCollapsed="Show More">This content is visible only when the component is expanded.</Expandable>
       </React.Fragment>
     );
   }